requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
gspread
google-auth
google-api-python-client
selenium
